GUI Added for Login
Login: Now passes the users dictionary to LoginGUI class and the GUI
takes care of logging in the user. It will return a valid user.
LoginGUI: All new. Uses GUI to get information from user. Many style
updates to come in the future on the GUI.
Maze: Changed framerate to something more fair on a faster computer :/