Fixed Login Loophole
Changed "login.py" to "start.py" because it does not login the user, it
merely acts as a main.
Added "self.__valid" attribute to loginGUI and if the password is not
correct, self.__valid is false and the user will not be added to
userOnline unless it is true.
Also prevented the user '' from registering.